+++ Classes/UVArticle.h
@@ -11,12 +11,14 @@
 @interface UVArticle : UVBaseModel {
     NSString *question;
     NSString *answerHTML;
+    NSInteger articleId;
 }
 
 + getInstantAnswers:(NSString *)query delegate:(id)delegate;
 
 @property (nonatomic, retain) NSString *question;
 @property (nonatomic, retain) NSString *answerHTML;
+@property (assign) NSInteger articleId;
 
 - (id)initWithDictionary:(NSDictionary *)dict;