+++ Classes/UVArticleViewController.m
@@ -7,6 +7,7 @@
 //
 
 #import "UVArticleViewController.h"
+#import "UVSession.h"
 
 @implementation UVArticleViewController
 
@@ -25,7 +26,6 @@
     [self hideExitButton];
     self.webView = [[[UIWebView alloc] initWithFrame:[self contentFrame]] autorelease];
     NSString *html = [NSString stringWithFormat:@"<html><head><link rel=\"stylesheet\" type=\"text/css\" href=\"http://cdn.uservoice.com/stylesheets/vendor/typeset.css\"/></head><body class=\"typeset\" style=\"font-family: sans-serif; margin: 1em\"><h3>%@</h3>%@</body></html>", article.question, article.answerHTML];
-    NSLog(@"%@", html);
     [self.webView loadHTMLString:html baseURL:nil];
     self.view = self.webView;
 }